Feat: 정원 api 연동 Feat: 화분 한번에 보기 api 연동

+++ wantPlant/app/src/main/java/com/example/wantplant/data/local/Pot.kt
@@ -1,7 +1,7 @@
 package com.example.wantplant.data.local
 
 data class Pot(
-    val potId: Int,
+    val potId: Long,
     val potName: String,
     val potTagColor: String,
     val proceed: Int,
